requirements.txt
# Weaviate クライアント
weaviate-client==4.9.4

# Ollama統合
langchain==0.3.12
langchain-community==0.3.12
langchain-ollama==0.2.1

# PDF処理
pypdf==5.1.0
pypdfium2==4.30.1

# テキスト処理
python-dotenv==1.0.1

# CLI
rich==13.9.4
typer==0.15.1

# RAGAS評価
ragas==0.4.3
datasets>=4.0.0
pandas==2.2.3
openai>=1.0.0  # RAGAS 0.4でOllamaをOpenAI互換APIとして使用
anthropic>=0.18.0  # Claude API for comparison
